==> rscripts/analysis/15-study1.qmd <==
---
title: "SEPTA --- Data analysis report"
subtitle: "Stockholm3 in a multi-ethnic cohort for prostate cancer detection (SEPTA): a prospective multicentered trial"
author: 
  - name: Andrea Discacciati
date: now
date-format: YYYY-MM-DD HH:mm
published-title: "Report compiled"
format:
  html:
    theme: default
    toc: true
    number-sections: true
    self-contained: true
    fontsize: medium
    anchor-sections: true
    toc-location: body
    grid:
      sidebar-width: 0px
      body-width: 1400px
      margin-width: 0px
      gutter-width: 0rem
execute:
  echo: false
editor_options: 
  chunk_output_type: console
params:
  save_tables: "false"
  time: "xxx"
---

```{r, include=FALSE}
source(here::here("rscripts", "analysis", "01-packages.R"))
source(here::here("rscripts", "analysis", "02-functions.R"))

#datetime_load <- "20230907_1201"
#latest dataset picked automatically
datetime_load <- tail(str_extract(sort(list.files(here::here("data", "derived"))), 
                                  "\\d{8}_\\d{4}"), 1)
datetime_load
septa <- readRDS(here::here("data", "derived", paste0(datetime_load, "-septa.rds"))) |> 
  make_test_pca_vars() |> 
  make_alternative_pca_vars()

septa_with_overall <- bind_rows(
  septa,
  septa |> mutate(race = "Overall")
) |> 
  mutate(race = factor(race, levels = make_factor_labels()$race_overall)) 

tables_list <- vector(mode = "list")
```

Derived data time stamp: ``r format(attr(septa, "datetime_created"), usetz = TRUE)``.

# Table 1
```{r}
tables_list$table_1 <- septa |>
  make_table1_vars() |> 
  select(race, age, psa, risk_score, dre_abnormal, fhpca, prevbx, five_ari, 
         risk_score, mri_performed, pirads_score, targeted_biopsy_performed, cb_cores_total, 
         cb_benign, cb_isup1, cb_isup2p, al_isup2p, cb_isup3p) |> 
  tbl_summary(by = race,
              missing = "ifany",
              missing_text = "Missing",
              #type = all_dichotomous() ~ "categorical",
              digits = list(psa ~ 1, cb_cores_total ~ 0)) |>  
  add_overall() |> 
  as_gt() |> 
  tab_options(table.font.size = px(14))

tables_list$table_1
```

```{r}
bloodtests <- c(`All` = "all", 
                `PSA>=3` = "psa_3p", 
                `PSA>=4` = "psa_4p", 
                `Stockholm3>=11`= "s3_11p", 
                `Stockholm3>=15` = "s3_15p")

bloodtests_table_2_psa4 <- bloodtests[c("PSA>=4", "Stockholm3>=11", "Stockholm3>=15")]
bloodtests_table_2_psa3 <- bloodtests[c("PSA>=3", "Stockholm3>=11", "Stockholm3>=15")]
bloodtests_table_4 <- bloodtests[c("All", "Stockholm3>=11", "Stockholm3>=15")]
```


# Table 2
relative Sensitivity, relative Specificity.

```{r}
table_2 <- split(septa_with_overall, septa_with_overall$race) |> 
  map(\(x) make_table_2(x, 
                        d = "cb_isup2p",
                        y1 = bloodtests_table_2_psa4[1], 
                        y2 = bloodtests_table_2_psa4[2],
                        y3 = bloodtests_table_2_psa4[3],
                        testlabels = names(bloodtests_table_2_psa4))) 
```

```{r}
# P-values for heterogeneity _for the main contrast Stockholm3≥15 versus PSA≥4_
contrast_matrix <- matrix(c(1, -1,  0,  0, 
                            1,  0, -1,  0, 
                            1,  0,  0, -1), 
                          ncol = 4, 
                          byrow = TRUE)

log.rel.sens_race <- log(sapply(table_2, `[[`, c("result", "y1y3", "sensitivity", "rel.sens")))[make_factor_labels()$race]
se.log.rel.sens_race  <- sapply(table_2, `[[`, c("result", "y1y3", "sensitivity", "se.log.rel.sens"))[make_factor_labels()$race]
chi2.rel.sens_race <- aod::wald.test(Sigma = diag(se.log.rel.sens_race^2), 
                                     b = log.rel.sens_race, 
                                     L = contrast_matrix)

log.rel.spec_race <- log(sapply(table_2, `[[`, c("result", "y1y3", "specificity", "rel.spec")))[make_factor_labels()$race]
se.log.rel.spec_race  <- sapply(table_2, `[[`, c("result", "y1y3", "specificity", "se.log.rel.spec"))[make_factor_labels()$race]
chi2.rel.spec_race <- aod::wald.test(Sigma = diag(se.log.rel.spec_race^2), 
                                     b = log.rel.spec_race, 
                                     L = contrast_matrix)
```

```{r}
tables_list$table_2 <- table_2 |> 
  map(`[[`, "out_df") |> 
  bind_rows(.id = "race") |> 
  gt_table_2() |> 
  tab_footnote(sprintf(
    "P-value for heterogeneity for Stockholm3 >=15 versus PSA >=4 ng/ml %.2g (relative Sensitivity), %.2g (relative Specificity).",
    chi2.rel.sens_race$result$chi2["P"], chi2.rel.spec_race$result$chi2["P"]
  ))

tables_list$table_2
```

P-values for heterogeneity _for the main contrast Stockholm3≥15 versus PSA≥4_:

P-value for heterogeneity (relative Sensitivity): `r format.pval(chi2.rel.sens_race$result$chi2["P"], digits = 2)` [ We had no evidence to reject the null hypothesis that relative Sensitivity was equal for the four race groups. / We had no evidence to conclude that relative Sensitivity was different across the four race groups. ] 

P-value for heterogeneity (relative Specificity): `r format.pval(chi2.rel.spec_race$result$chi2["P"], digits = 2)` [ We had evidence to reject the null hypothesis that relative Specificity was equal for the four race groups. / We had evidence to conclude that relative Specificity was different across the four race groups. ]


# Table 3
Sensitivity, Specificity, PPV, NPV.

```{r}
table_3 <- split(septa_with_overall, septa_with_overall$race) |> 
  map(\(x)
      imap(bloodtests, 
           \(y, tl) make_row_table_3(dat = x, d = "cb_isup2p", y = y, testlabel = tl)
      )
  )

tables_list$table_3 <- table_3 |> 
  map(
    \(r) map(r, `[[`, "out_df") |> bind_rows()
  ) |> 
  bind_rows(.id = "race") |> 
  gt_table_3()

tables_list$table_3
```




# Figure 1
```{r}
#| fig-width: 8
#| fig-height: 7

data_panel_rel.sens <- data.frame(
  race = names(table_2),
  rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["sensitivity"]]["rel.sens"])),
  lcl.rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["sensitivity"]]["lcl.rel.sens"])),
  ucl.rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["sensitivity"]]["ucl.rel.sens"]))
) |> 
  mutate(race = factor(race, levels = rev(make_factor_labels()$race_overall)))

panel_rel.sens <- plot_figure_1(dat = data_panel_rel.sens,
                                title = "Relative Sensitivity (95% CI)",
                                panel = "rsens")


data_panel_rel.spec <- data.frame(
  race = names(table_2),
  rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["specificity"]]["rel.spec"])),
  lcl.rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["specificity"]]["lcl.rel.spec"])),
  ucl.rel = sapply(table_2, \(x) unname(x[["result"]][["y1y3"]][["specificity"]]["ucl.rel.spec"]))
) |> 
  mutate(race = factor(race, levels = rev(make_factor_labels()$race_overall)))

panel_rel.spec <- plot_figure_1(dat = data_panel_rel.spec,
                                title = "Relative Specificity (95% CI)",
                                panel = "rspec")

patchwork::wrap_plots(panel_rel.sens,
                      patchwork::plot_spacer(),
                      panel_rel.spec,
                      patchwork::plot_spacer(),
                      nrow = 4,
                      heights = c(4, 1, 4, 1))
```


# Subgroup analyses

## Table 2
```{r}
septa_strata <- make_strata_vars(septa)
strata_vars <- names(septa_strata)[startsWith(names(septa_strata), "strata_")]
strata_vars <- setNames(strata_vars, labelled::var_label(septa_strata[, strata_vars]))

table_2_sa <- map(strata_vars, \(s)
                  map(split(septa_strata, septa_strata[[s]]), \(dat)
                      make_table_2(dat, 
                                   d = "cb_isup2p",
                                   y1 = bloodtests_table_2_psa4[1], 
                                   y2 = bloodtests_table_2_psa4[2],
                                   y3 = bloodtests_table_2_psa4[3],
                                   testlabels = names(bloodtests_table_2_psa4))$out_df   
                  ) |> bind_rows(.id = "stratum")
) |> bind_rows(.id = "sa")

tables_list$table_2_sa <- table_2_sa |> 
  gt_table_2()

tables_list$table_2_sa
```

## Table 3

```{r}
table_3_sa <- map(
  strata_vars, \(s)
  map(split(septa_strata, septa_strata[[s]]), \(dat)
      imap(bloodtests, 
           \(y, tl) make_row_table_3(dat = dat, d = "cb_isup2p", y = y, testlabel = tl)$out_df
      ) |> bind_rows()
  ) |> bind_rows(.id = "stratum")
) |> bind_rows(.id = "sa")

tables_list$table_3_sa <- table_3_sa |> 
  gt_table_3()

tables_list$table_3_sa
```

## Table 4
```{r}
strata2_vars <- names(septa_strata)[startsWith(names(septa_strata), "strata2_")]
strata2_vars <- setNames(strata2_vars, labelled::var_label(septa_strata[, strata2_vars]))


table_4_sa <- bind_rows(
  # overall
  map(strata2_vars[3], \(s)
      map(split(septa_strata, septa_strata[[s]]), \(dat)
          imap(bloodtests_table_4, 
               \(y, tl) make_row_table_3(dat = dat, d = "cb_isup2p", y = y, testlabel = tl)$out_df
          ) |> bind_rows()
      ) |> bind_rows(.id = "stratum")
  ) |> bind_rows() |> 
    mutate(sa = "Overall",
           sa1 = "Overall", .before = everything()),
  # by race
  map(strata2_vars[1:2], \(s1) 
      map(split(septa_strata, septa_strata[[s1]]), \(s2) 
          map(split(s2, s2[[strata2_vars[3]]]), \(dat)
              imap(bloodtests_table_4, 
                   \(y, tl) make_row_table_3(dat = dat, d = "cb_isup2p", y = y, testlabel = tl)$out_df
              ) |> bind_rows()
          ) |> bind_rows(.id = "stratum")
      ) |> bind_rows(.id = "sa1")
  ) |> bind_rows(.id = "sa")
)

tables_list$table_4_sa <- table_4_sa |> 
  gt_table_4()

tables_list$table_4_sa
```



# AUC

```{r}
septa_roc <- septa_with_overall |> 
  pivot_longer(cols = c(psa, risk_score)) |> 
  mutate(text = ifelse(name == "psa", "PSA (ng/ml)", "Stockholm3"))
```


## Combined biopsy

Outcome: ISUP2+ on combined biopsy (SBx+TBx)

```{r message = FALSE, warning = FALSE}
auc_roc_test <- make_table_auc(septa_roc, "cb_isup2p")

tables_list$auc_roc_test <- auc_roc_test |> 
  gt_table_auc()

tables_list$auc_roc_test
```

## Combined biopsy, alternative outcome definition

Outcome: unfavourable intermediate or higher cancers

```{r message = FALSE, warning = FALSE}
auc_roc_test_iu <- make_table_auc(septa_roc, "al_isup2p")

tables_list$auc_roc_test_iu <- auc_roc_test_iu |> 
  gt_table_auc()

tables_list$auc_roc_test_iu
```

## Standard biopsy in men without targeted biopses

Outcome: ISUP2+ on standard biopsy  

```{r message = FALSE, warning = FALSE}
auc_roc_test_notb <- make_table_auc(filter(septa_roc, targeted_biopsy_performed == "No"), "sb_isup2p")

auc_roc_test_notb |> 
  gt_table_auc()
```

# ROC curves

## Combined biopsy

Outcome: ISUP2+ on combined biopsy (SBx+TBx)

```{r}
#| fig-width: 12
#| fig-height: 10

roc_plots <- plot_roc_curves(cb_isup2p, septa_roc, auc_roc_test) +
  labs(title = "ROC curves for the Stockholm3 test and PSA (ng/ml) (ISUP>=2 cancers)")

roc_plots
```


## Combined biopsy, alternative outcome definition

Outcome: unfavourable intermediate or higher cancers

```{r}
#| fig-width: 12
#| fig-height: 10

roc_plots_iu <- plot_roc_curves(al_isup2p, septa_roc, auc_roc_test_iu) +
  labs(title = "ROC curves for the Stockholm3 test and PSA (ng/ml) (Unfavourable Intermediate or higher cancers)")

roc_plots_iu
```

## Standard biopsy in men without targeted biopses

Outcome: ISUP2+ on standard biopsy  

```{r}
#| fig-width: 12
#| fig-height: 10

roc_plots_notb <- plot_roc_curves(sb_isup2p, septa_roc, auc_roc_test_notb) +
  labs(title = "ROC curves for the Stockholm3 test and PSA (ng/ml) (ISUP>=2 cancers in men without targeted biopsies)")

roc_plots_notb
```


# Calibration curves

## Combined biopsy

Outcome: ISUP2+ on combined biopsy (SBx+TBx)

```{r}
#| fig-width: 12
#| fig-height: 10

calibration_plots <- imap(split(septa_with_overall, septa_with_overall$race),
                          \(x, y)   plot_calibration_curves(y = cb_isup2p,
                                                            p = (risk_score/100), # from S3M score to probability
                                                            data = x,
                                                            xlab = "Predicted ISUP >=2 probability",
                                                            ylab = "Observed ISUP >=2 probability",
                                                            title = y)$plot$lowess)

patchwork::wrap_plots(calibration_plots, nrow = 2) +
  patchwork::plot_annotation(
    title = "Calibration plots for the Stockholm3 test (ISUP>=2 cancers)"
  )
```

## Combined biopsy, alternative outcome definition

Outcome: unfavourable intermediate or higher cancers

```{r}
#| fig-width: 12
#| fig-height: 10

calibration_plots_iu <- imap(split(septa_with_overall, septa_with_overall$race),
                             \(x, y)   plot_calibration_curves(y = al_isup2p,
                                                               p = (risk_score/100), # from S3M score to probability
                                                               data = x,
                                                               xlab = "Predicted Unfav. Interm. or higher probability",
                                                               ylab = "Observed Unfav. Interm. or higher probability",
                                                               title = y)$plot$lowess)

patchwork::wrap_plots(calibration_plots_iu, nrow = 2) +
  patchwork::plot_annotation(
    title = "Calibration plots for the Stockholm3 test (Unfavourable Intermediate or higher cancers)"
  )
```


## Standard biopsy in men without targeted biopses

Outcome: ISUP2+ on standard biopsy  

```{r}
#| fig-width: 12
#| fig-height: 10
septa_with_overall_notb <- filter(septa_with_overall, 
                                  targeted_biopsy_performed == "No")

calibration_plots_notb <- imap(split(septa_with_overall_notb, septa_with_overall_notb$race),
                               \(x, y)   plot_calibration_curves(y = sb_isup2p,
                                                                 p = (risk_score/100), # from S3M score to probability
                                                                 data = x,
                                                                 xlab = "Predicted ISUP >=2 probability",
                                                                 ylab = "Observed ISUP >=2 probability",
                                                                 title = y)$plot$lowess)

patchwork::wrap_plots(calibration_plots_notb, nrow = 2) +
  patchwork::plot_annotation(
    title = "Calibration plots for the Stockholm3 test (in men without targeted biopsies)"
  )
```

# Decision curves

Outcome: ISUP2+ on combined biopsy (SBx+TBx)

```{r warning=FALSE, message=FALSE}
#| fig-width: 12
#| fig-height: 10
dca_plots <- imap(split(septa_with_overall, septa_with_overall$race),
                  \(x, y) plot_dca(x, y))

patchwork::wrap_plots(dca_plots, nrow = 2) +
  patchwork::plot_annotation(
    title = "Decision curve analysis for the Stockholm3 test and PSA (ng/ml)"
  ) +
  patchwork::plot_layout(guides = 'collect') &
  theme(legend.position='bottom')
```


# Stockholm3 threshold and specificity at fixed sensitivity ≥90%

Outcome: ISUP2+ on combined biopsy (SBx+TBx) 

```{r}
tables_list$s3m_fixed_sens_90 <- map(
  split(septa_with_overall, septa_with_overall$race), calculate_s3m_threshold_at_fixed_sens
) |> bind_rows(.id = "race") |> 
  select(-FPF) |> 
  mutate(across(c(TPF, TNF), format_percent)) |> 
  gt() |> 
  cols_label(
    race = "Race",
    TNF = "Specificity",
    TPF = "Sensitivity",
    c = "Stockholm3 threshold (>=)"
  ) |> 
  tab_style(
    style = cell_text(weight = "bold"),
    locations = list(cells_column_labels(),
                     cells_column_spanners())
  ) |> 
  tab_options(table.font.size = px(14))

tables_list$s3m_fixed_sens_90
```



# Table year/recruitment/targeted biopsy

```{r}
tables_list$table_year_recruitment_tbx <- full_join(
  septa_strata |> 
    count(biopsy_year, strata_septa, .drop = FALSE) |> 
    pivot_wider(names_from = "strata_septa", values_from = "n",names_prefix = "septa_"),
  septa_strata |> 
    count(biopsy_year, targeted_biopsy_performed, .drop = FALSE) |> 
    pivot_wider(names_from = "targeted_biopsy_performed", values_from = "n", names_prefix = "tbx_") |> 
    select(biopsy_year, tbx_Yes),
  by = "biopsy_year"
) |> 
  mutate(Total = septa_No + septa_Yes,
         septa_Yes = paste0(septa_Yes, " (", format_percent(septa_Yes/Total), ")"),
         septa_No  = paste0(septa_No,  " (", format_percent(septa_No/Total),  ")"),
         tbx_Yes   = paste0(tbx_Yes,   " (", format_percent(tbx_Yes/Total),   ")")) |> 
  gt() |> 
  cols_label(
    biopsy_year = "Year of biopsy",
    septa_No = "No, n (%)",
    septa_Yes = "Yes, n (%)",
    tbx_Yes = "MRI-Targeted biopsies performed, n (%)",
    Total = "Total, n"
  ) |> 
  tab_spanner(
    label = "Men specifically recruited for SEPTA", columns = c("septa_No", "septa_Yes")
  ) |> 
  tab_style(
    style = cell_text(weight = "bold"),
    locations = list(cells_column_labels(),
                     cells_column_spanners())
  ) |> 
  tab_options(table.font.size = px(14))

tables_list$table_year_recruitment_tbx
```

# Cancers missed/detected by race and blood test

```{r}
cancers_missed_detected <- septa_with_overall |> 
  pivot_longer(cols = unname(bloodtests[-1])) |> 
  select(name, value, combined_grade_group, race) |> 
  mutate(value = factor(value, labels = make_factor_labels()$ny),
         name = factor(name, levels = bloodtests[-1],
                       labels = names(bloodtests[-1]))) |> 
  split(~ race) |> 
  imap(
    \(s, n) tbl_strata(
      select(s, -race),
      strata = name,
      .tbl_fun =
        \(x) tbl_summary(x, by = value,
                         label = list(combined_grade_group ~ "Combined biopsy"),
                         all_categorical() ~ "{n}") |> 
        modify_header(label = paste0("**", n, "**")),
      .header = "**{strata}**"
    ) |> 
      as_gt() |> 
      tab_options(table.font.size = px(14))
  )
```

```{r results='asis'}
cancers_missed_detected
```

```{r}
names(cancers_missed_detected) <- paste0("cancers_missed_detected_", names(cancers_missed_detected))
tables_list <- c(tables_list, cancers_missed_detected)
```



# Alternative biopsy outcome definition


Cross-classification Combined biopsy GG versus Alternative biopsy outcome definition

```{r}
with(septa,
     addmargins(tablem(alternative_grade_group, combined_grade_group)))
```


## Table 2
```{r}
table_2_alt <- split(septa_with_overall, septa_with_overall$race) |> 
  map(\(x) make_table_2(x, 
                        d = "al_isup2p",
                        y1 = bloodtests_table_2_psa4[1], 
                        y2 = bloodtests_table_2_psa4[2],
                        y3 = bloodtests_table_2_psa4[3],
                        testlabels = names(bloodtests_table_2_psa4))) 

tables_list$table_2_alt <- table_2_alt |> 
  map(`[[`, "out_df") |> 
  bind_rows(.id = "race") |> 
  gt_table_2_alt()

tables_list$table_2_alt
```

## Table 3


```{r}
table_3_alt <- split(septa_with_overall, septa_with_overall$race) |> 
  map(\(x)
      imap(bloodtests, 
           \(y, tl) make_row_table_3(dat = x, 
                                     d = "al_isup2p", 
                                     y = y, 
                                     testlabel = tl)
      )
  )

tables_list$table_3_alt <- table_3_alt |> 
  map(
    \(r) map(r, `[[`, "out_df") |> bind_rows()
  ) |> 
  bind_rows(.id = "race") |> 
  gt_table_3_alt()

tables_list$table_3_alt
```

# Table 2 for Stockholm3≥11 and Stockholm≥15 vs PSA≥3 ng/ml

Outcome: ISUP2+ on combined biopsy (SBx+TBx)

```{r}
table_2_psa3 <- split(septa_with_overall, septa_with_overall$race) |> 
  map(\(x) make_table_2(x, 
                        d = "cb_isup2p",
                        y1 = bloodtests_table_2_psa3[1], 
                        y2 = bloodtests_table_2_psa3[2],
                        y3 = bloodtests_table_2_psa3[3],
                        testlabels = names(bloodtests_table_2_psa3))) 

tables_list$table_2_psa3 <- table_2_psa3 |> 
  map(`[[`, "out_df") |> 
  bind_rows(.id = "race") |> 
  gt_table_2()

tables_list$table_2_psa3
```

# Relative PPV and NPV for Stockholm3≥{11, 15} vs PSA≥4 ng/ml

Point estimates and 95% joint confidence region

```{r}
#| fig-width: 12.5
#| fig-height: 10

rpv_data <- list(
  `11` = map(table_2, \(x) pv.rpv(x$data_table$y1y2)),
  `15` = map(table_2, \(x) pv.rpv(x$data_table$y1y3))
)

#imap(rpv_data, \(x, y) plot_rpv_ellipse(x, y))
plot_rpv_ellipse_overlay(rpv_data)

gt_table_rpv(rpv_data)
```

# Table 1 by SEPTA recruitment
```{r}
tables_list$table_1_biobank <- septa |>
  make_strata_vars() |> 
  make_table1_vars() |> 
  select(race, age, psa, risk_score, dre_abnormal, fhpca, prevbx, five_ari, 
         risk_score, mri_performed, pirads_score, targeted_biopsy_performed, cb_cores_total, 
         cb_benign, cb_isup1, cb_isup2p, al_isup2p, cb_isup3p, strata_septa) |> 
  tbl_summary(by = strata_septa,
              missing = "ifany",
              missing_text = "Missing",
              #type = all_dichotomous() ~ "categorical",
              digits = list(psa ~ 1, cb_cores_total ~ 0)) |>  
  add_overall() |> 
  modify_spanning_header(all_stat_cols() ~ "**Men specifically recruited for SEPTA**") |> 
  as_gt() |> 
  tab_options(table.font.size = px(14))

tables_list$table_1_biobank
```

# Additional descriptive statistics for PSA and Age

```{r}
tables_list$table1_additional_stats <- septa |> 
  select(race, age, psa, race) |> 
  tbl_summary(by = race,
              missing = "ifany",
              missing_text = "Missing",
              type = all_continuous() ~ "continuous2",
              statistic = all_continuous() ~ c("{median} ({p25}, {p75})", 
                                               "{mean}", 
                                               "{min}-{max}"),
              digits = list(psa ~ 1)) |> 
  add_overall()  |> 
  as_gt() |> 
  tab_options(table.font.size = px(14))

tables_list$table1_additional_stats
```


# Two-by-two biomarker tables 

```{r warning=FALSE, results='asis'}
bloodtest_crosstable <- 
  map(bloodtests[2:3], \(p)
      map(bloodtests[4:5], \(s) {
        c.kappa <- function(x) {
          # x must be prop.table
          p.o <- sum(diag(x))
          p.yes <- rowSums(x)[2] * colSums(x)[2]
          p.no  <- rowSums(x)[1] * colSums(x)[1]
          p.e <- p.yes + p.no
          kappa <- (p.o-p.e)/(1-p.e)
          unname(kappa)
        }
        
        df <- septa |> 
          mutate(across(c(p, s), \(x) 
                        factor(x, levels = c(0, 1), labels = make_factor_labels()$ny))) |> 
          labelled::set_variable_labels(.labels = setNames(as.list(make_var_labels()$label),
                                                           make_var_labels()$var), 
                                        .strict = FALSE)
        
        kappa <- c.kappa(prop.table(table(df[[p]], df[[s]])))
        footnote <- paste0("Cohen's kappa coefficient: ", 
                           sprintf("%4.2f", kappa), ".")
        
        df |> 
          tbl_cross(row = p, 
                    col = s, 
                    percent = "cell") |> 
          modify_header(label = ".") |> # see https://github.com/ddsjoberg/gtsummary/issues/1613
          as_gt() |> 
          tab_footnote(footnote) 
      }
      
      )
  ) |> 
  list_flatten()
```

```{r results='asis'}
bloodtest_crosstable
```

```{r}
names(bloodtest_crosstable) <- paste0("bloodtest_crosstable_", 
                                      names(bloodtest_crosstable))
tables_list <- c(tables_list, bloodtest_crosstable)
```


```{r, include = FALSE}
#| fig-width: 12.1
#| fig-height: 8.5

# Extra: Spearman's rank correlation coefficient
spearman <- split(septa_with_overall, ~ race) |> 
  map_dfr(~ cor.test(.$psa, .$risk_score, method = "spearman", exact = FALSE)$estimate, .id = "race") |> 
  mutate(race = factor(race, levels = c("Overall", levels(septa$race)))) 

ggplot(septa_with_overall, aes(x = psa, y = risk_score/100, color = race)) +
  geom_point(data = select(septa, -race), 
             aes(x = psa, y = risk_score/100), 
             inherit.aes = FALSE, color = grey(0.8)) + 
  geom_point(show.legend = FALSE) +
  geom_label(data = spearman, 
             aes(8, 0.012,
                 label = paste0("Spearman's correlation coefficient = ", sprintf("%5.3f", rho))),
             inherit.aes = FALSE, 
             label.size = 0) +
  facet_wrap(~ race, nrow = 2, scales = "free") +
  scale_x_log10(limits = c(0.1, 2000), labels = scales::label_number()) +
  scale_y_continuous(trans = "logit", breaks = c(1, 3, seq(10, 90, 10), 97, 99)/100, limits = c(0.01, 0.99),
                     labels = \(x) round(x*100, 0)) +
  scale_color_manual(values = c("Black", RColorBrewer::brewer.pal(4, "Set1"))) +
  theme_minimal(base_size = 12) +
  theme(strip.text = element_text(size = 13)) +
  labs(x = "PSA (ng/ml)",
       y = "Stockholm3 Risk Score",
       caption = "x-axis is in the log10 scale. y-axis is in the logit scale.") 
```

# PIRADS score by combined biopsy outcome

```{r}
xtabs(~ pirads_score + combined_grade_group, data = septa) |> 
  addmargins()
```

# Concordance between PSA and PSA for Stockholm3

```{r}
pearson <- with(septa, cor.test(psa, psa_a3p))
```

Pearson's correlation coefficient between PSA and PSA for Stockholm3: `r sprintf("%5.3f", pearson$estimate)`.

```{r}
#| fig-width: 6
#| fig-height: 6
ggplot(septa, aes(x = psa, y = psa_a3p)) +
  geom_point(pch = 1, cex = 2, alpha = 0.5) +
  geom_abline(col = "red", lwd = 0.8) + 
  scale_x_log10(limits = c(0.1, 2000), labels = scales::label_number()) +
  scale_y_log10(limits = c(0.1, 2000), labels = scales::label_number()) +
  annotate("text", x = 100, y = 0.15, label = paste0("Pearson's correlation coefficient = ", sprintf("%5.3f", pearson$estimate))) +
  theme_minimal(base_size = 12) +
  labs(x = "PSA (ng/ml)",
       y = "PSA (ng/ml) for Stockholm3",
       caption = "Red line is the identity line.") +
  theme(aspect.ratio = 1)
```

Extra: MA plot for PSA versus PSA for Stockholm3 

```{r}
#| fig-width: 6
#| fig-height: 6
ggplot(septa, aes(x = 0.5*(log10(psa_a3p)+log10(psa)), y = log10(psa_a3p)-log10(psa))) +
  geom_point(pch = 1, cex = 2, alpha = 0.5) +
  geom_hline(yintercept = 0, col = "red", lwd = 0.7) +
  geom_smooth(method = "loess", se = FALSE, col = "deepskyblue", lwd = 0.7) +
  scale_y_continuous(limits = c(-2.5, 2.5), n.breaks = 10) +
  scale_x_continuous(n.breaks = 10) +
  theme_minimal(base_size = 12) +
  labs(x = "A = log10(PSA for Stockholm3 * PSA)/2",
       y = "M = log10(PSA for Stockholm3 / PSA)",
       caption = "Red line is referent line at M=0 (ie, ratio=1). Blue line is loess smoother.") +
  theme(aspect.ratio = 1)
```

(MA plot is a 45-degrees--rotated and stretched version of the standard scatter 
plot with axes in the log-scale. See code in qmd file.)

```{r include = FALSE, eval = FALSE}
#| fig-width: 6
#| fig-height: 6
dta <- with(septa, cbind(
  log10(psa_a3p),
  log10(psa),
  0.5*(log10(psa * psa_a3p)), # A transform
  log10(psa / psa_a3p))       # M transform
)

# Rotation matrix
R <- matrix(c(cos(pi/4), -sin(pi/4), 
              sin(pi/4), cos(pi/4)),
         nrow = 2, ncol = 2, byrow = TRUE)

# Stretch matrix
S <- matrix(c(1/sqrt(2), 0, 
              0, sqrt(2)),
            nrow = 2, ncol = 2, byrow = TRUE)

head(dta[, 1:2] %*% R %*% S)
head(dta[, 3:4])

o <- dta[, 1:2] %*% R
oo <- o %*% S
par(pty = "s")
plot(dta[, 1], dta[, 2], ylim = c(-2.5, 4), xlim = c(-2.5, 4), pch = 20, cex = 0.5, col = "gray")
points(o[, 1], o[, 2], col = "tomato", pch = 20, cex = 0.5)
points(oo[, 1], oo[, 2], col = "gold3", pch = 20, cex = 0.5)
```


```{r}
if(params$save_tables == "true") {
  filename_save <- here::here("output", paste0(params$time, "-septa_tables_study1.docx"))
                         
  gt_group(.list = tables_list) |> 
    gtsave(filename_save)
}
```


# Session info

```{r}
sessionInfo()
```
